\name{transcribedFeatures}
\alias{transcribedFeatures}
\non_function{}
\title{A character vector of GFF feature names of things that can be transcribed}
\description{A character vector of GFF feature names of things that can be transcribed}
\usage{
data(transcribedFeatures)
}
\format{A character vector}
\author{W. Huber <huber@ebi.ac.uk>}
\examples{
data(transcribedFeatures)
}
\keyword{datasets}
