\name{showDens}
\alias{showDens}
\title{Plot function for more than one density}
\description{
  Plot function for more than one density
}
\usage{
showDens(z, breaks, col, ylab = "", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{z}{List of numeric vectors for computing histograms for}
  \item{breaks}{breaks of the histogram}
  \item{col}{colour of the histograms}
  \item{ylab}{y-axis label}
  \item{\dots}{futher arguments passed on to plot}
}
\details{
  ...
}
\value{
  returns scale factor
}
\author{Wolfgang Huber \email{huber@ebi.ac.uk}}
\seealso{\code{\link{hist}}}
\examples{
}
\keyword{internal}% at least one, from doc/KEYWORDS
